renamed locustfile towards good naming
